chore: split out router events into its own module
